update-03

+++ src/models/admin/ActivityLog.js
@@ -4,10 +4,10 @@ const { sequelize } = require("../../config/db");
 const ActivityLog = sequelize.define(
   "ActivityLog",
   {
-    id: { type: DataTypes.INTEGER.UNSIGNED, primaryKey: true, autoIncrement: true },
+    id: { type: DataTypes.BIGINT.UNSIGNED, primaryKey: true, autoIncrement: true },
 
     adminId: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       allowNull: false,
       references: {
         model: "admins",
