+++ src/models/admin/booking/BookingPayment.js
@@ -5,14 +5,14 @@ const BookingPayment = sequelize.define(
   "BookingPayment",
   {
     id: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       primaryKey: true,
       autoIncrement: true,
     },
 
     // FK → booking.id
     bookingId: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       allowNull: false,
       references: {
         model: "booking",
