+++ src/models/admin/booking/CancelBooking.js
@@ -5,13 +5,13 @@ const CancelBooking = sequelize.define(
   "CancelBooking",
   {
     id: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       autoIncrement: true,
       primaryKey: true,
     },
 
     bookingId: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       allowNull: false,
     },
     bookingType: {
