+++ src/models/admin/sessionPlan/SessionPlanGroup.js
@@ -5,7 +5,7 @@ const SessionPlanGroup = sequelize.define(
   "SessionPlanGroup",
   {
     id: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       autoIncrement: true,
       primaryKey: true,
     },
@@ -53,7 +53,7 @@ const SessionPlanGroup = sequelize.define(
       allowNull: false,
     },
     sortOrder: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       allowNull: false,
       defaultValue: 0,
     },
@@ -68,7 +68,7 @@ const SessionPlanGroup = sequelize.define(
     },
     // ✅ Foreign key to admins table for creation
     createdBy: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       allowNull: false,
       references: {
         model: "admins",
@@ -86,7 +86,7 @@ const SessionPlanGroup = sequelize.define(
 
     // ✅ Foreign key to admins table for deletion
     deletedBy: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       allowNull: true,
       references: {
         model: "admins",
