+++ src/models/admin/venue/venue.js
@@ -6,7 +6,7 @@ const Venue = sequelize.define(
   "Venue",
   {
     id: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       autoIncrement: true,
       primaryKey: true,
     },
@@ -41,7 +41,7 @@ const Venue = sequelize.define(
     // },
     // ✅ Use FK to PaymentGroups
     paymentGroupId: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       allowNull: true,
       references: {
         model: "payment_groups", // 👈 your table name
@@ -83,7 +83,7 @@ const Venue = sequelize.define(
     },
     // ✅ Foreign key to admins table for creation
     createdBy: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       allowNull: true,
       references: {
         model: "admins",
@@ -102,7 +102,7 @@ const Venue = sequelize.define(
 
     // ✅ Foreign key to admins table for deletion
     deletedBy: {
-      type: DataTypes.INTEGER.UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       allowNull: true,
       references: {
         model: "admins",
